fact-check-bot/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.1
pydantic>=2.10.0
openai>=1.30.1
requests>=2.31.0
transformers>=4.41.2
torch>=2.9.0
sentencepiece>=0.2.0
spacy>=3.8.0
asyncpraw>=7.7.0
feedparser>=6.0.0
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
numpy>=1.24.0
newspaper3k>=0.2.8
lxml_html_clean>=0.4.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
